quiplash/requirements.txt
azure-core==1.36.0
azure-cosmos==4.14.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
